moved socket to its own file left a pub use in lib.rs so existing imports will continue to function
